Trying to Add Lives
I have Added a lives variable to the DestroyByContact script.
The lives are stored in the enemies, I would think it would be easier
to store them in the TooBee script but I wasn’t sure how to wire that
up right. I also created a new text view right below Score to indicate
the number of lives. I changed the enemy prefab to have 3 lives I also
tried to wire that up to the GUI lives text but I couldn’t do that
unless it an instance of it was in the main hierarchy.  I think the key
to stopping the game from completely resetting will have to be
somewhere inside the GameController’s gameOver method that is called
when the player collides with an enemy.