cleaned up code and prepped code for tomorrows meeting
good news! I was able to clean up the code pretty quickly. there was a
lot of mess in there but it looks a ton better now!
also, the issue with lives looks like we will easily solve it if all
goes well, I put some comments in the code to help guide us