added a new coin prefab and a script
I still have more work to do but as it stands right now I’ve got a coin
prefab with a script called InvulnerabilityCollision attached to it.
when TooBee runs into the coin the coin disappears.

We may end up changing the name of the script or removing all together
and moving the code to another script like DestroyByContact.